[refactor] #52 : 경로 변경
