+++ src/feature/my/model/index.ts
@@ -1,0 +1,1 @@
+export * from "./Password.type";
